Разделяет utils.js на отдельные файлы и добавляет функцию remove
